fix: Trocando o atributo balance para BigDecimal

+++ src/main/java/com/gabriel/desafiopicpay/controller/dto/response/WalletResponse.java
@@ -1,7 +1,8 @@
 package com.gabriel.desafiopicpay.controller.dto.response;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 public record WalletResponse(UUID id,
-                             Integer balance) {
+                             BigDecimal balance) {
 }
